moved messages dialogues and posts out of local components and placed within App
